Move multipart download classes to proper package
