+++ docs/src/main/asciidoc/resteasy-reactive.adoc
@@ -587,7 +587,7 @@ import javax.ws.rs.Path;
 import javax.ws.rs.Produces;
 import javax.ws.rs.core.MediaType;
 
-import org.jboss.resteasy.reactive.server.core.multipart.MultipartFormDataOutput;
+import org.jboss.resteasy.reactive.server.multipart.MultipartFormDataOutput;
 
 @Path("multipart")
 public class Endpoint {
