+++ independent-projects/resteasy-reactive/server/processor/src/main/java/org/jboss/resteasy/reactive/server/processor/generation/multipart/FormDataOutputMapperGenerator.java
@@ -19,10 +19,10 @@ import org.jboss.jandex.Type;
 import org.jboss.logging.Logger;
 import org.jboss.resteasy.reactive.common.processor.JavaBeanUtil;
 import org.jboss.resteasy.reactive.common.processor.ResteasyReactiveDotNames;
-import org.jboss.resteasy.reactive.server.core.multipart.MultipartFormDataOutput;
 import org.jboss.resteasy.reactive.server.core.multipart.MultipartMessageBodyWriter;
 import org.jboss.resteasy.reactive.server.core.multipart.MultipartOutputInjectionTarget;
-import org.jboss.resteasy.reactive.server.core.multipart.PartItem;
+import org.jboss.resteasy.reactive.server.multipart.MultipartFormDataOutput;
+import org.jboss.resteasy.reactive.server.multipart.PartItem;
 
 import io.quarkus.gizmo.AssignableResultHandle;
 import io.quarkus.gizmo.ClassCreator;
