+++ independent-projects/resteasy-reactive/server/runtime/src/main/java/org/jboss/resteasy/reactive/server/core/multipart/MultipartOutputInjectionTarget.java
@@ -1,5 +1,7 @@
 package org.jboss.resteasy.reactive.server.core.multipart;
 
+import org.jboss.resteasy.reactive.server.multipart.MultipartFormDataOutput;
+
 public interface MultipartOutputInjectionTarget {
     MultipartFormDataOutput mapFrom(Object pojo);
 }
